replace category plugin by contrib/category-lite plugin. fix #528
